examples/plivo-chatbot/requirements.txt
pipecat-ai[noisereduce,elevenlabs,cartesia,openai,anthropic,silero,deepgram] @ git+https://github.com/pipecat-ai/pipecat.git@main
fastapi
uvicorn
python-dotenv
loguru

flask
duckduckgo-search
phidata
